// src/app/models/parametros/proponente.model.ts
export class ProponenteModel{
    id?: number;
    primer_nombre?: string;
    otros_nombres?: string;
    primer_apellido?: string;
    segundo_apellido?: string;
    documento?: number;
    fecha_nacimiento?: string;
    email?: string;
    celular?: string;
    image?: string;
    image_public_id?: string;
    id_tipo_vinculacion?: number;
}